+++ x/configuration/genesis.go
@@ -6,6 +6,7 @@ import (
 
 	sdk "github.com/cosmos/cosmos-sdk/types"
 	"github.com/iov-one/iovns/x/configuration/types"
+	domain_types "github.com/iov-one/iovns/x/domain/types"
 )
 
 // GenesisState is used to unmarshal the genesis state
@@ -60,23 +61,48 @@ func ValidateGenesis(data GenesisState) error {
 // DefaultGenesisState returns the default genesis state
 // TODO this needs to be updated, although it will be imported from iovns chain
 func DefaultGenesisState() GenesisState {
+	// get owner
+	owner, err := sdk.AccAddressFromBech32("star1kxqay5tndu3w08ps5c27pkrksnqqts0zxeprzx")
+	if err != nil {
+		panic("invalid default owner provided")
+	}
+	// set default configs
+	config := types.Config{
+		Configurer:             owner,
+		ValidDomain:            "^(.*?)?",
+		ValidName:              "^(.*?)?",
+		ValidBlockchainID:      "^(.*?)?",
+		ValidBlockchainAddress: "^(.*?)?",
+		DomainRenew:            86400,
+	}
+	// set fees
+	fees := types.NewFees()
+	defFee := sdk.NewCoin("iov", sdk.NewInt(10))
+	// add domain fees
+	fees.UpsertDefaultFees(&domain_types.MsgRegisterDomain{}, defFee)
+	fees.UpsertDefaultFees(&domain_types.MsgAddAccountCertificates{}, defFee)
+	fees.UpsertDefaultFees(&domain_types.MsgDeleteAccountCertificate{}, defFee)
+	fees.UpsertDefaultFees(&domain_types.MsgDeleteDomain{}, defFee)
+	fees.UpsertDefaultFees(&domain_types.MsgDeleteAccount{}, defFee)
+	fees.UpsertDefaultFees(&domain_types.MsgFlushDomain{}, defFee)
+	fees.UpsertDefaultFees(&domain_types.MsgRegisterAccount{}, defFee)
+	fees.UpsertDefaultFees(&domain_types.MsgRenewAccount{}, defFee)
+	fees.UpsertDefaultFees(&domain_types.MsgRenewDomain{}, defFee)
+	fees.UpsertDefaultFees(&domain_types.MsgReplaceAccountTargets{}, defFee)
+	fees.UpsertDefaultFees(&domain_types.MsgTransferAccount{}, defFee)
+	fees.UpsertDefaultFees(&domain_types.MsgTransferDomain{}, defFee)
+	fees.UpsertDefaultFees(&domain_types.MsgSetAccountMetadata{}, defFee)
+	// return genesis
 	return GenesisState{
-		Config: types.Config{
-			Configurer:             sdk.AccAddress{},
-			ValidDomain:            "^(.*?)?",
-			ValidName:              "^(.*?)?",
-			ValidBlockchainID:      "^(.*?)?",
-			ValidBlockchainAddress: "^(.*?)?",
-			DomainRenew:            86400,
-		},
-
-		Fees: types.NewFees(),
+		Config: config,
+		Fees:   fees,
 	}
 }
 
 // InitGenesis sets the initial state of the configuration module
 func InitGenesis(ctx sdk.Context, k Keeper, data GenesisState) {
 	k.SetConfig(ctx, data.Config)
+	k.SetFees(ctx, data.Fees)
 }
 
 // ExportGenesis saves the state of the configuration module