+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## HEAD
 
+## v0.2.5
+
 - iovnscli: fix has-superuser bool flag bug
 - iovnsd: fix duplicate blockchain targets ID
 - remove flush domain feature